Character factory now sets correct finite state machine
