--- Assets/AIBrain.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 15a47a883a78e3d4f90576bf5b85218b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
