--- Assets/AIBrain/PlayerBrain.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 869363281f44a3043aff2727f9b68644
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
