+++ Assets/Abilities/Spells.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dbf6b8537431fb14792aedb66953331a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
