--- Assets/PlayerTurnStateBehavior.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e6083e15bd4415e4eb5106879b9f17bb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
