+++ Assets/Resources/Ai.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 15a47a883a78e3d4f90576bf5b85218b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
