+++ Assets/Resources/Ai/StrikerBrain.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4b4d1bb25fb331d419830e4259cc85bf
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
