+++ Assets/Scripts/Battlefield/BrainManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0a183cd50fa54344488323417f6fd811
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
