--- Assets/Scripts/Battlefield/CreatureScripts/CreatureBase.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f8de4b27c7559b148baec73b65de674c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
