--- Assets/Scripts/Battlefield/StateMachineBehaviors.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 09080f98691131847ad856edfd8974ff
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
